UDP association: implement stateful packet handling
The SOCKS5 UDP associations should, apparently, handle replies by sending them
back to the originating client port. Thus, we need to keep track of the remote
addresses that the client communicates with, so we know which client port to
reply to. I had earlier misinterpreted the RFC as saying that DEST in the
request specified where replies should be sent through - DEST can legally be all
zeroes. We now ignore DEST completely (it's not mandatory to use it for access
control purposes).

This should fix #3 properly, providing OpenVPN support over UDP through the
proxy.